pyproject.toml
# View options here: https://python-poetry.org/docs/configuration
[tool.poetry]
# Change this to match the directory your project is in
# Poetry will install your project as an editable package
name = "transcription_bot"
version = "0.1"
description = "Whisper-based Telegram transcription bot"
authors = ["extrange <29305375+extrange@users.noreply.github.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
cryptg = "^0.4.0"
uvloop = "^0.20.0"
pillow = "^10.4.0"
hachoir = "^3.3.0"
pysubs2 = "^1.7.3"
replicate = "^0.31.0"
humanize = "^4.10.0"
minio = "^7.2.7"
telethon = "^1.36.0"
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pyright = "^1.1.375"
ruff = "^0.5.7"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
coverage = "^7.6.1"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
# pre-commit: make pyright use the venv
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
# https://docs.pytest.org/en/latest/explanation/goodpractices.html#which-import-mode
addopts = ["--import-mode=importlib"]
asyncio_mode = "auto"